requirements.txt
botcity-framework-core>=1.0.0,<2.0
botcity-framework-web>=0.8.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0
pandas
openpyxl
gdown
python-environ
loguru
environ